+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/CourseRestController.java
@@ -18,4 +18,9 @@ public class CourseRestController {
     public Course createCourse(@RequestBody CourseDTO courseDTO, @PathVariable Long schoolId){
         return courseController.createCourse(courseDTO, schoolId);
     }
+
+    @PutMapping("/{courseId}")
+    public Course updateCourseName(@PathVariable Long courseId, @RequestBody CourseDTO courseDTO){
+        return courseController.updateCourseName(courseId, courseDTO);
+    }
 }