+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/CourseController.java
@@ -24,4 +24,10 @@ public class CourseController {
         return courseService.createCourse(courseDTO, school);
 
     }
+
+    public Course updateCourseName(Long courseId, CourseDTO courseDTO) {
+        Course course = courseService.getCourseById(courseId);
+        course.setCourseName(courseDTO.getCourseName());
+        return courseService.saveCourse(course);
+    }
 }