+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/CourseService.java
@@ -20,4 +20,12 @@ public class CourseService {
         courseRepository.save(course);
         return course;
     }
+
+    public Course getCourseById(Long courseId) {
+        return courseRepository.findById(courseId).orElse(null);
+    }
+
+    public Course saveCourse(Course course) {
+        return courseRepository.save(course);
+    }
 }